# add fastbuild option to ./build/release.sh and use it in make quick-release