--- a/hack/lib/golang.sh
+++ b/hack/lib/golang.sh
@@ -47,11 +47,45 @@ kube::golang::server_targets() {
 readonly KUBE_SERVER_TARGETS=($(kube::golang::server_targets))
 readonly KUBE_SERVER_BINARIES=("${KUBE_SERVER_TARGETS[@]##*/}")
 
-# The server platform we are building on.
-readonly KUBE_SERVER_PLATFORMS=(
-  linux/amd64
-  linux/arm
-)
+if [[ "${KUBE_FASTBUILD:-}" == "true" ]]; then
+  readonly KUBE_SERVER_PLATFORMS=(linux/amd64)
+  readonly KUBE_TEST_PLATFORMS=(linux/amd64)
+  if [[ "${KUBE_BUILDER_OS:-}" == "darwin"* ]]; then
+    readonly KUBE_CLIENT_PLATFORMS=(
+      darwin/amd64
+      linux/amd64
+    )
+  else
+    readonly KUBE_CLIENT_PLATFORMS=(linux/amd64)
+  fi
+else
+
+  # The server platform we are building on.
+  readonly KUBE_SERVER_PLATFORMS=(
+    linux/amd64
+    linux/arm
+  )
+
+  # If we update this we should also update the set of golang compilers we build
+  # in 'build/build-image/cross/Dockerfile'. However, it's only a bit faster since go 1.5, not mandatory
+  readonly KUBE_CLIENT_PLATFORMS=(
+    linux/amd64
+    linux/386
+    linux/arm
+    darwin/amd64
+    darwin/386
+    windows/amd64
+    windows/386
+  )
+
+  # Which platforms we should compile test targets for. Not all client platforms need these tests
+  readonly KUBE_TEST_PLATFORMS=(
+    linux/amd64
+    darwin/amd64
+    windows/amd64
+    linux/arm
+  )
+fi
 
 # The set of client targets that we are building for all platforms
 readonly KUBE_CLIENT_TARGETS=(
@@ -59,18 +93,6 @@ readonly KUBE_CLIENT_TARGETS=(
 )
 readonly KUBE_CLIENT_BINARIES=("${KUBE_CLIENT_TARGETS[@]##*/}")
 readonly KUBE_CLIENT_BINARIES_WIN=("${KUBE_CLIENT_BINARIES[@]/%/.exe}")
-
-# If we update this we should also update the set of golang compilers we build
-# in 'build/build-image/cross/Dockerfile'. However, it's only a bit faster since go 1.5, not mandatory 
-readonly KUBE_CLIENT_PLATFORMS=(
-  linux/amd64
-  linux/386
-  linux/arm
-  darwin/amd64
-  darwin/386
-  windows/amd64
-  windows/386
-)
 
 # The set of test targets that we are building for all platforms
 kube::golang::test_targets() {
@@ -106,14 +128,6 @@ readonly KUBE_TEST_PORTABLE=(
   hack/get-build.sh
   hack/ginkgo-e2e.sh
   hack/lib
-)
-
-# Which platforms we should compile test targets for. Not all client platforms need these tests
-readonly KUBE_TEST_PLATFORMS=(
-  linux/amd64
-  darwin/amd64
-  windows/amd64
-  linux/arm
 )
 
 # Gigabytes desired for parallel platform builds. 11 is fairly