Don't build lot's of architectures.

timing done on my 32gb/12core workstation:

With:
```
make quick-release  73.37s user 6.64s system 71% cpu 1:51.14 total
```

Without:
```
make quick-release  132.21s user 10.77s system 53% cpu 4:26.45 total
```

cc @david-mcmahon @thockin @smarterclayton

TODO: test kube-up on osx

I declare war on slow builds.