@smarterclayton 2016-02-19
This isn't you I don't think, but may be related to the new build image:

```
KUBE_RELEASE_RUN_TESTS=n KUBE_FASTBUILD=true build/release.sh
+++ [0218 22:15:35] Verifying Prerequisites....
+++ [0218 22:15:39] Building Docker image kube-build:build-7799c106a6.
+++ Docker build command failed for kube-build:build-7799c106a6

Sending build context to Docker daemon 32.04 MB
Sending build context to Docker daemon 
Step 0 : FROM gcr.io/google_containers/kube-cross:v1.4.2-1
Trying to pull repository gcr.io/google_containers/kube-cross ... a91de1002859: Layer already being pulled by another client. Waiting.
a91de1002859: Download complete
failed
time="2016-02-18T22:16:16-05:00" level=info msg="could not find image: no such id: a91de1002859284af6df4de100bb7f38d7e0b4c5f73b85cbd3ee9e9efe352e6a" 

To retry manually, run:

docker build -t kube-build:build-7799c106a6 --pull=false /Users/clayton/projects/go/src/k8s.io/kubernetes/_output/images/kube-build:build-7799c106a6
```