@smarterclayton 2016-02-19
Got a clean run on a mac.