@mikedanese 2016-02-19
@smarterclayton I think you may have run into https://github.com/kubernetes/kubernetes/issues/20904