@smarterclayton 2016-02-19
Maybe, I had no problems after the image completed pulling.