@luxas 2016-02-19
/cc @luxas

I already made it a bit faster when I removed `addon` builds for `non-linux-amd64` and splitted client targets to client and test.

Do you think we need test targets for `darwin/amd64` and `windows/amd64` or should we remove them too? I may send a fix for that. (for faster cross-builds)