@zmerlynn 2016-02-23
@k8s-bot e2e test this, issue #IGNORE (no test results)